Refactored components and improved file structure
