+++ src/App.js
@@ -38,7 +38,7 @@ import {
     faFreeCodeCamp,
   } from '@fortawesome/free-brands-svg-icons'
   import { WelcomeModal } from './components/WelcomeModal';
-  import { WorkTile } from './components/WorkTile';
+  import { TileBox } from './components/Tiles';
 
 class App extends Component {
   render() {
@@ -78,7 +78,9 @@ class App extends Component {
           </Hero>
         <WelcomeModal />
         </Section>
-        <MainContent />
+        <Section>
+          <TileBox />
+        </Section>
         <EmailForm />
         <Footer>
           <Container>
@@ -112,66 +114,6 @@ class EmailForm extends Component {
           </Level>
         </Container>
       </Section>
-    )
-  }
-}
-
-class MainContent extends Component {
-  render() {
-    return (
-      <Section>
-          <TileBox />
-      </Section>
-    )
-  }
-}
-
-class TileBox extends Component {
-  render() {
-    return (      
-      <Tile kind="ancestor" vertical>
-        <Tile size={12} kind="parent">
-              <GoalTile />
-              <ExperienceTile />
-        </Tile>
-        <Tile size={12} kind="parent">
-            <EducationTile />
-            <WorkTile/>
-        </Tile>
-      </Tile>
-    )
-  }
-}
-
-class GoalTile extends Component {
-  render() {
-    return (
-      <Tile kind="child" size={4} notification color="link">
-        <Heading>MY GOAL</Heading>
-        <p>To collaborate with a great team to create quality, mobile-first software experiences that make people feel good.<br /> <br /> Ultimately, I want to make the world better every day.<br /><br /><span role="img" aria-label="rocketship emoji">🚀</span></p>
-      </Tile>
-    )
-  }
-}
-
-class ExperienceTile extends Component {
-  render() {
-    return (
-      <Tile kind="child" notification>
-        <Heading>DEV EXPERIENCE</Heading>
-        <p><a href="http://www.dunecoffee.com">dunecoffee.com</a> -<strong> css style upkeep + animations + custom theming + wholesale portal</strong><br /><br /> portfolio - react website<br /><br /> hawaiicjc - js interactivity<br /><br /> github - demonstration of creativity, entrepreneurial spirit</p>
-      </Tile>
-    )
-  }
-}
-
-class EducationTile extends Component {
-  render() {
-    return (
-      <Tile kind="child"  size={8} notification color="primary">
-        <Heading>EDUCATION</Heading>
-        <p>Udacity - Grow with Google Codecademy - Intro to JS freecodecamp Always a lifelong learner, dying to learn more and stay sharp</p>
-      </Tile>
     )
   }
 }
